+++ src/main/java/kr/jbnu/se/std/Weaponduck.java
@@ -13,7 +13,7 @@ public class Weaponduck {
     public int hp;
     public BufferedImage weadckImg;
 
-    public Weaponduck(int x, int y, int speed, int score, int hp, BufferedImage weadckImg){
+    public Weaponduck(int x, int y, float speed, int score, int hp, BufferedImage weadckImg){
         this.x = x;
         this.y = y;
         this.hp = Game.getlvdata().bosshp/3;
@@ -25,17 +25,17 @@ public class Weaponduck {
     public void Draw(Graphics2D g2d){g2d.drawImage(weadckImg, x, y, null);}
 
     public static class Smgduck extends Weaponduck{
-        public Smgduck(int x, int y, int speed, int score, int hp, BufferedImage weadckImg) {
+        public Smgduck(int x, int y, float speed, int score, int hp, BufferedImage weadckImg) {
             super(x, y, speed, score, hp, weadckImg);
         }
     }
     public static class Rifduck extends Weaponduck{
-        public Rifduck(int x, int y, int speed, int score, int hp, BufferedImage weadckImg) {
+        public Rifduck(int x, int y, float speed, int score, int hp, BufferedImage weadckImg) {
             super(x, y, speed, score, hp, weadckImg);
         }
     }
     public static class Odinduck extends Weaponduck{
-        public Odinduck(int x, int y, int speed, int score, int hp, BufferedImage weadckImg) {
+        public Odinduck(int x, int y, float speed, int score, int hp, BufferedImage weadckImg) {
             super(x, y, speed, score, hp, weadckImg);
         }
     }
